[fix] 상세 정보 조회시, 메이커 정보 없는 경우 테스트

+++ src/main/java/com/snapfit/main/post/infra/persistence/PostRepositoryImpl.java
@@ -250,7 +250,7 @@ public class PostRepositoryImpl implements PostRepository {
     }
 
     @Override
-    public Mono<PageResult<Post>> findByMaker(int limit, int offset, int makerId) {
+    public Mono<PageResult<Post>> findByMaker(int limit, int offset, long makerId) {
         return databaseClient.sql("""
                         SELECT
                             p.id AS post_id,
